ポモドーロ機能のリファクタリングと一時停止・再開機能の準備
